Separate base server app from notebookap
