+++ scripts/jupyter-server
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+from notebook.serverapp import main
+
+if __name__ == '__main__':
+    main()
